Make `extendsProfile` recursive
