+++ src/api/index.ts
@@ -1,3 +1,4 @@
+export * from './profile'
 export * from './base'
 export * from './compiler'
 export * from './editor'
